Fix 'undefined references' linker error
Windows compilation of the latest code of fab-agon-emulator fails with linker errors about undefined references to few fabgl::VGADirectController methods (drawArc / fillSegment / fillSector).

Looks like the problem is that those functions are not implemented as proper class members in VGADirectControllers.cpp, but as free functions outside of the class. For whatever reasons, this is not a problem for the Linux build, but on Windows it caused build failures.