increase the stack size used by “primitives” task
this is the task that handles the drawing queue

this fixes a bug - calls to `Canvas::fillSector` were causing crashes, with stack traces that indicated a stack/heap corruption

the underlying `GenericBitmappedDisplayController::genericFillSector` function that handles the `FillSector` primitive is _complicated_, and uses a _lot_ of local variables.  a very rough estimate is that it may use more than the 1200 bytes that were being assigned for this task’s stack.  it’s no wonder we were seeing crashes

the new stack size of 3000 bytes an educated safe guess, and is not fine-tuned.  (it’s lower than the first guess of 5000 bytes).  this level seems to be OK and programs function reliably7 - we can still use all of our existing screen modes, and there seems to be no adverse effects